Add KEEP_ON_FAILURE flag to e2e kind tests (#832)
Using this flag allows debugging the failed test on the environment,
instead of destroying it.
The default behavior is still to clean up the cluster when failing.

Fixes: https://github.com/istio-ecosystem/sail-operator/issues/675

+++ tests/e2e/multicontrolplane/multi_control_plane_test.go
@@ -170,6 +170,10 @@ spec:
 	})
 
 	AfterAll(func() {
+		if CurrentSpecReport().Failed() && keepOnFailure {
+			return
+		}
+
 		By("Cleaning up the application namespaces")
 		Expect(k.DeleteNamespace(appNamespace1, appNamespace2a, appNamespace2b)).To(Succeed())
 
@@ -188,9 +192,15 @@ spec:
 	})
 
 	AfterAll(func() {
-		if CurrentSpecReport().Failed() && !debugInfoLogged {
-			common.LogDebugInfo(common.MultiControlPlane, k)
-			debugInfoLogged = true
+		if CurrentSpecReport().Failed() {
+			if !debugInfoLogged {
+				common.LogDebugInfo(common.MultiControlPlane, k)
+				debugInfoLogged = true
+			}
+
+			if keepOnFailure {
+				return
+			}
 		}
 
 		if skipDeploy {
